Fix two Sprokit python-related link errors
Fix two issues with sprokit_pipeline linking to sprokit_python_util.
First, this link was happening unconditionally, despite that the latter
library is not built (or needed) if Python support is not enabled.
Second, the latter library was being declared later, such that it is not
known when the former's link dependencies are declared.

This fixes build failures introduced by #341 (139962dc5bfb).