Introduce indirect range adapter
Create a range adapter to apply a level of indirection to a range during
iteration. This effectively suppresses dereferencing of the iterator, so
that the user sees the iterators rather than the values, while still
being able to use the more terse and convenient syntax of a range-based
for. Also, fix some minor lingering issues with creating a mutating
range adapter.