Prevent using range adapters on temporaries
Add r-value reference overloads to prevent using range adapters on
temporaries, as doing so causes UB because the input range goes out of
scope before the loop can execute.

It appears possible to fix this, at least for copyable (or movable)
ranges, by having the range adapter in such cases take a copy of the
input range. However, this requires overloading on r-value references,
which is broken on GCC 4.8 which (for now) we still support. Thus,
unfortunately, the best we can do for now is avoid such code by making
it not compile on non-broken compilers.