track_oracle/KPF activity: populate actor_intervals; split out apply()
This commit splits loading KPF activities into read() and apply() methods.
It also adds the actor_interval field.

The read() method loads the file and populates all the schema fields except
for actor_tracks, which requires geometry. This allows the KPF activity file
to be read and analyzed WITHOUT requiring the geometry to be present.

The apply() method takes the activity "tracks" and a set of reference geometry
tracks, and clones the appropriate reference geometry into the activity's
actor_tracks fields.