[stab] Add basic stabilization support to SRNN tracker for motion features
The key aspect of this approach is to add a member "ref_point" to the
track_state class that tracks bbox_center in a stabilized coordinate
system.  This is then used to compute motion features etc. instead of
bbox_center.

SRNNTracker._step gets a slightly annoying piece of logic to handle
missing homography input and breaks in a semi-sensible manner (namely
assuming identity).

It would be good to make all coordinates relative to the current frame
to mitigate problems with drift or straight-up bad homographies.