Don't return detector name by reference
Change detected_object::detector_name to return by value, rather than by
reference. Although this is a little more expensive, returning by
reference is dangerous, and the performance argument is hard to justify
when we are returning stuff like the notes and key points, which are
potentially much larger and more complicated data sets, by value.