Use more explicit namespaces in ads
In the header and implementation files for adapter_data_set,
the namespace specifier sprokit:: has been replaced with ::sprokit:: in a few
method signatures and implementations. MSVC would have trouble differentiating
between the kwiver::sprokit namespace used in bindings, and the ::sprokit:: namespace
used in the header and implementation files.

Suggested-by: Brad King <brad.king@kitware.com>