Pretty-print vector_3d in tests
Add Google Test pretty-printer for kwiver::vector_3d (a.k.a.
Eigen::Vector3d). This makes it possible for humans to read the output
when a vector comparison fails.

(I'm not entirely clear why this is needed, as Eigen matrices are
supposed to have stream operators, which Google Test is, in theory,
supposed to use. I suspect something may have changed at some point, and
all matrices are affected, but we can revisit that later, as needed.
This came up while trying to get the new PROJ implementation working, as
the geo_point test uses vector_3d.)