Overhaul PROJ arrow, use PROJ 5 API
Overhaul the PROJ arrow to use the PROJ 5/6 APIs. We need to do this
eventually as the PROJ 4 API is deprecated and must be opted into for
PROJ 6.x, and is removed entirely by PROJ 8.x.

This has some good bits, such as theoretically being thread safe, which
the old implementation probably was not. It also has a bunch of bad
bits, such as the number of cached projections now being O(N²) rather
than O(N) (for N being the number of coordinate systems "in use"),
because the PROJ 5 API is, frankly, harder to use and in some ways
deliberately obtuse. There is also an unfortunate quantity of code that
depends on whether we are targeting PROJ 5 or PROJ 6, as there are some
nasty behavior changes between the two that make them incompatible,
though the situation isn't quite as bad as moving from 4 to 5.