Fix Python build
Several places are failing entirely to add the includes for PyBind11,
and those places that do are adding them via include_directories.
Instead, link the pybind11::pybind11 target in all places that need
PyBind11.

This fixes trying to build with a PyBind11 that is not in an include
directory that is already being used for some other reason.