Enhance adapter_data_set API with value[_or]
Add value and value_or methods to adapter_data_set. The former replaces
get_port_data as a more "usual" API name, while the latter expands on
these to allow the user to request a value while providing a fallback in
case no such value is available, rather than throwing an exception. This
can make it easier to defensively get data, particularly for pointer
data types which have an obvious "not available" value.