Fix OCV inpainter modifying its inputs
Modify ocv::inpaint to avoid modifying the input images. Doing so is
unexpected and potentially dangerous, especially when the algorithm is
being used within a pipeline. Moreover, the manner in which we were
creating the output image container, which ultimately shared memory with
the input, was creating a divorced reference count, which could result
in the output becoming invalid when the input image container reference
count hits zero.