Always guess install prefix if not specified
Modify functions in config_block_io that determine or use search paths
to always attempt to guess the install prefix from the location of the
executing binary if no prefix is specified. (Note that this excludes the
overload which takes the application and library/KWIVER prefixes
separately.)

This addresses the need for every user of these APIs to replicate this
logic. (However, those users which were already doing so are left alone,
as they have superior mechanisms of determining the prefix.)