"Fix" keyframe selection
Modify initialize_cameras_landmarks to skip keyframe down-selection and
force use of all frames for initial reconstruction if the number of
input frames is smaller than the maximum number of keyframes to use.
This is required because the down-selection process will otherwise
down-select to an empty set if only a few frames have tracks, which will
be the case if the input is not video, or if only a few frames have
features when using manually annotated feature tracks.