Create video metadata filter
Create an algorithm that wraps a video input and applies a
transformation to the metadata. Right now, this is implemented via a
pure virtual function, meaning that users need to subclass the algorithm
to use it. (This will likely be replaced, later, with a proper
algorithm, but for now we don't have such an algorithm yet.)