Actually use unique_ptr in metadata map
Add a small helper header for the couple of places in which we want to
use std::unique_ptr in a std::map. Change the default to actually using
unique_ptr rather than the work-around, which should only be needed on
broken STL implementations. Fix the numerous bugs (mostly needing to
avoid unnecessary copies) due to actually using unique_ptr.