Removed "future" cmake modules for an "old" version
Current CMake version is beyond 3.10, so including "future" modules for
that version no longer makes sense.
Additionally, the message to indicate that a future being included has
been moved to within the inner if-statement that actually adds the
directory to the module path.
Added a status message to indicate if a futures directory exists but is
ignored given the version comparison with the current CMake version.