Fixed a broken bundle adjust test
The test was checking if all cameras are shared after bundle adjustment
when forcing shared cameras, but the problem was that the test cameras
were shared to begin with, so the test would pass if no sharing change
was made.  This change forces the test cameras to use unique intrincs
from the start to test of the bundle adjustment forces them to be common
again.