Finish computing matches before modifying the feature tracks
If not, there's at least one issue due to threading, namely that the
call to merge_tracks can ultimately call erase on
simple_track_set_implementation::data_, which can cause the call to
simple_track_set_implementation::tracks, which copies data_ and is
indirectly made by match_tracks, to copy data_ while elements are
actively being removed from it.  This manifests as empty shared_ptrs
in the return value of tracks that are dereferenced in
track_set_implementation::active_tracks to cause a segfault.

Commit 17c7936e6b3514fc9ce677f74eb2b283cfe36ebf appears to have broken
this.