Update kwiver-setup-python
A major fix here is updating the determination of the python site
package. This was incorrectly adding a "local/" prefix for some
platforms (linux), splitting where python library stuff was going.
I'm not sure if this is going to work for windows as I don't have a test
machine for that...

This also cleans up cruft where we had been maintaining using the
backwards compatible "PYTHON_*" prefixed variables throughout the codebase
despite the `file_package( Python )` having long since moved to a "Python_*"
syntax.